+++ R/http.R
@@ -23,7 +23,7 @@ TWIT_post <- function(token, api, params = NULL, body = NULL, ..., host = "api.t
 
 TWIT_method <- function(method, token, api, 
                         params = NULL, 
-                        host = "api.twiter.com",
+                        host = "api.twitter.com",
                         retryonratelimit = NULL,
                         verbose = TRUE,
                         ...) {
@@ -34,23 +34,22 @@ TWIT_method <- function(method, token, api,
   token <- check_token(token)
   url <- paste0("https://", host, api, ".json")
   
-  repeat({
-    resp <- switch(method,
-      GET = httr::GET(url, query = params, token, ...),
-      POST = httr::POST(url, query = params, token, ...),
-      stop("Unsupported method", call. = FALSE)
-    )
-    
-    switch(resp_type(resp),
-      ok = break,
-      rate_limit = handle_rate_limit(
-        resp, api, 
-        retryonratelimit = retryonratelimit,
-        verbose = verbose
-      ),
-      error = handle_error(resp)
-    )
-  })
+  resp <- switch(method,
+                 GET = httr::GET(url, query = params, token, ...),
+                 POST = httr::POST(url, query = params, token, ...),
+                 stop("Unsupported method", call. = FALSE)
+  )
+  
+  switch(resp_type(resp),
+         ok = NULL,
+         protected = handle_protected(resp, params),
+         rate_limit = handle_rate_limit(
+           resp, api, 
+           retryonratelimit = retryonratelimit,
+           verbose = verbose
+         ),
+         error = handle_error(resp)
+  )
 
   resp
 }
@@ -87,7 +86,7 @@ TWIT_method <- function(method, token, api,
 #' @param since_id Supply a vector of ids or a data frame of previous results to 
 #'   find tweets **newer** than `since_id`.
 #' @param retryonratelimit If `TRUE`, and a rate limit is exhausted, will wait
-#'   until it refreshes. Most twitter rate limits refresh every 15 minutes.
+#'   until it refreshes. Most Twitter rate limits refresh every 15 minutes.
 #'   If `FALSE`, and the rate limit is exceeded, the function will terminate
 #'   early with a warning; you'll still get back all results received up to 
 #'   that point. The default value, `NULL`, consults the option 
@@ -188,7 +187,6 @@ TWIT_paginate_cursor <- function(token, api, params,
                                  retryonratelimit = NULL,
                                  verbose = TRUE) {
   params$count <- page_size
-  
   cursor <- next_cursor(cursor)
   if (identical(cursor, "0")) {
     # Last request retrieved all available results
@@ -219,19 +217,28 @@ TWIT_paginate_cursor <- function(token, api, params,
     )
 
     if (is_rate_limit(json)) {
-      warn_early_term(json, 
-        hint = paste0("Set `cursor = '", cursor, "' to continue."),
-        hint_if = !identical(cursor, "-1")
-      )
+      if (!is.null(retryonratelimit)){
+        warn_early_term(json, 
+                        hint = paste0("Set `cursor = '", cursor, "' to continue."),
+                        hint_if = !identical(cursor, "-1")
+        )
+      }
       break
     }
 
     results[[i]] <- json
-    cursor <- json$next_cursor_str
+    if (any(grepl("next_cursor", names(json)))) {
+      cursor <- ifelse(!is.null(json$next_cursor_str), 
+                       json$next_cursor_str, 
+                       json$next_cursor)
+    } else {
+      # If next_cursor is missing there are no message within the last 30 days
+      cursor <- "0" 
+    }
     n_seen <- n_seen + length(get_id(json))
     i <- i + 1
-
-    if (identical(cursor, "0") || n_seen >= n) {
+    empty_response <- !is.null(json$events) && length(json$events) == 0
+    if (identical(cursor, "0") || n_seen >= n || empty_response) {
       break
     }
     
@@ -300,6 +307,8 @@ resp_type <- function(resp) {
   x <- resp$status_code
   if (x == 429) {
     "rate_limit"
+  } else if (x == 401) {
+    "protected"
   } else if (x >= 400) {
     "error"
   } else {
@@ -354,11 +363,19 @@ warn_early_term <- function(cnd, hint, hint_if) {
 # https://developer.twitter.com/en/support/twitter-api/error-troubleshooting
 handle_error <- function(x) {
   json <- from_js(x)
-  stop(
-    "Twitter API failed [", x$status_code, "]\n",
-    paste0(" * ", json$errors$message, " (", json$errors$code, ")"),
-    call. = FALSE
-  )
+  stop("Twitter API failed [", x$status_code, "]\n",
+       paste0(" * ", json$errors$message, " (", json$errors$code, ")"),
+       call. = FALSE)
+}
+
+handle_protected <- function(resp, params) {
+  handle_codes(resp)
+  
+  if (any(c("screen_name", "user_id") %in% names(params))) {
+    account <- params$screen_name
+    if (is.null(account)) account <- params$user_id
+    warning("Skipping unauthorized account: ", account, call. = FALSE)
+  }
 }
 
 check_status <- function(x, api) {
@@ -378,5 +395,21 @@ check_token <- function(token = NULL) {
     httr::add_headers(Authorization = paste0("Bearer ", token$token))
   } else {
     abort("`token` is not a valid access token")
+  }
+}
+
+# Function for responses that might be errors
+# Depending on the internal error code it is provided to the users as a warning or error
+handle_codes <- function(x) {
+  if ("errors" %in% names(httr::content(x))) {
+    errors <- httr::content(x)$errors[[1]]
+    for (e in seq_len(max(lengths(errors)))) {
+      funct <- switch(as.character(errors$code[e]),
+                      "89" = stop,
+                      warning)
+      funct(paste0(errors$message[e], " (", errors$code[e], ")"),
+           call. = FALSE)
+      
+    }
   }
 }