+++ R/lists_subscriptions.R
@@ -1,4 +1,5 @@
-#' Get list subscriptions of a given user.
+#' Get list subscriptions of a given user but does not include the user's own 
+#' lists.
 #'
 #' @inheritParams TWIT_paginate_cursor
 #' @inheritParams get_timeline
@@ -15,6 +16,7 @@
 #' }
 #'
 #' @family lists
+#' @references <https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/create-manage-lists/api-reference/get-lists-subscriptions>
 #' @export
 lists_subscriptions <- function(user,
                                 n = 20,