+++ R/ts_plot.R
@@ -34,7 +34,7 @@
 #'   n = 3000
 #' )
 #'
-#' ## examine all twitter activity using weekly intervals
+#' ## examine all Twitter activity using weekly intervals
 #' ts_plot(tmls, "weeks")
 #'
 #' ## group by screen name and plot each time series
@@ -129,7 +129,7 @@ ts_data_ <- function(data, by = "days", trim = 0L, tz = "UTC") {
   ## reformat time var
   .unit <- parse_unit(by)
   ## adjust to desired tz
-  data[[dtvar]] <- as.POSIXct(format(data[[dtvar]], tz = "UTC"), tz = tz)
+  data[[dtvar]] <- convert_tz(data[[dtvar]], tz = tz)
   data[[dtvar]] <- round_time(data[[dtvar]], by, tz)
   ## get unique values of time in series
   dtm <- unique(