+++ R/users.R
@@ -41,6 +41,7 @@ lookup_users <- function(users, parse = TRUE, token = NULL,
 
   if (parse) {
     results <- users_with_tweets(results)
+    results$created_at <- format_date(results$created_at)
   }
   
   results