+++ tests/testthat/test-lists_subscriptions.R
@@ -1,7 +1,10 @@
 test_that("lists_subscriptions returns lists data frame", {
-  x <- lists_subscriptions("kearneymw")
+  
+  vcr::use_cassette("lists_subscriptions1", {
+    x <- lists_subscriptions("kearneymw")
+  })
 
   expect_true(is.data.frame(x))
   expect_true("list_id" %in% names(x))
-  expect_gt(nrow(x), 1)
+  expect_true(nrow(x) >= 1)
 })