+++ tests/testthat/test-mentions.R
@@ -1,5 +1,8 @@
 test_that("mentions returns tweets data", {
-  suppressMessages(x <- get_mentions())
+  
+  vcr::use_cassette("mentions", {
+    suppressMessages(x <- get_mentions())
+  })
 
   expect_s3_class(x, "data.frame")
   expect_gt(nrow(x), 0)