+++ tests/testthat/test-tweet_threading.R
@@ -1,5 +1,8 @@
 test_that("tweet_threading works", {
-  tw <- lookup_tweets('1084143184664510470')
-  tw_thread <- tweet_threading(tw)
+  
+  vcr::use_cassette("tweet_threading", {
+    tw <- lookup_tweets('1461776330584956929')
+    tw_thread <- tweet_threading(tw)
+  })
   expect_s3_class(tw_thread, "data.frame")
 })